refactor: helper functions in game logic
