+++ src/components/Game/GameLogic.ts
@@ -1,7 +1,8 @@
-import {IAllPositions, IPositionSchema, TDirections} from '../../interfaces/api';
+import {IAllPlayers, IAllPositions, IPositionSchema, TDirections} from '../../interfaces/api';
 import {getRandomColumn, getRandomRow} from '../utils';
+import {COLUMNS, ROWS} from '../../consts';
 
-export function getOccupiedByPLayers(positions: IAllPositions, rowOrCol: ('row' | 'col')) {
+export function getOccupiedByPLayers(positions: IAllPositions, rowOrCol: ('row' | 'col')): number[] {
     //todo can be a set
     return Object.values(positions).reduce((prev: (number)[], currentPositions) => {
         const currentRows = currentPositions.position.map((position: IPositionSchema) => position[rowOrCol]);
@@ -9,12 +10,12 @@ export function getOccupiedByPLayers(positions: IAllPositions, rowOrCol: ('row' 
     }, []);
 }
 
-export function getOccupiedByFood(foods: IPositionSchema[], rowOrCol: ('row' | 'col')) {
+export function getOccupiedByFood(foods: IPositionSchema[], rowOrCol: ('row' | 'col')): number[] {
     // todo can be a set
     return foods.map((food) => food[rowOrCol]);
 }
 
-export function getUnoccupiedPosition(positions: IAllPositions, foods: IPositionSchema[]) {
+export function getUnoccupiedPosition(positions: IAllPositions, foods: IPositionSchema[]): IPositionSchema {
     // check that new random position is not clashing with existing players or food
     let randomRow = getRandomRow();
     const occupiedPlayerRows = getOccupiedByPLayers(positions, 'row');
@@ -38,6 +39,13 @@ export function getUnoccupiedPosition(positions: IAllPositions, foods: IPosition
     };
 }
 
+export function getAllColors(players: IAllPlayers): string[] {
+    return Object.values(players).reduce((acc: string[], player) => {
+        acc.push(player.color);
+        return acc;
+    }, []);
+}
+
 export function getUpdatedFood(foods: IPositionSchema[], eatenFood: IPositionSchema[]): IPositionSchema[] {
     const updatedFood: IPositionSchema[] = [];
 
@@ -52,7 +60,7 @@ export function getUpdatedFood(foods: IPositionSchema[], eatenFood: IPositionSch
     return updatedFood;
 }
 
-export function updateDirection(previousDirection: TDirections, newDirection: TDirections) {
+export function updateDirection(previousDirection: TDirections, newDirection: TDirections): TDirections {
 
     let updatedDirection = newDirection; // new direction by default
 
@@ -63,4 +71,25 @@ export function updateDirection(previousDirection: TDirections, newDirection: TD
     if (previousDirection === 'LEFT' && updatedDirection === 'RIGHT') updatedDirection = 'LEFT';
 
     return updatedDirection;
+}
+
+export function getNewHead(currentHead: IPositionSchema, updatedDirection: TDirections): IPositionSchema {
+    const newHead: IPositionSchema = {...currentHead};
+
+    switch (updatedDirection) {
+    case 'UP':
+        newHead.row = currentHead.row === 0 ? ROWS - 1 : currentHead.row - 1;
+        break;
+    case 'DOWN':
+        newHead.row = currentHead.row === ROWS - 1 ? 0 : currentHead.row + 1;
+        break;
+    case 'LEFT':
+        newHead.col = currentHead.col === 0 ? COLUMNS - 1 : currentHead.col - 1;
+        break;
+    case 'RIGHT':
+        newHead.col = currentHead.col === COLUMNS - 1 ? 0 : currentHead.col + 1;
+        break;
+    }
+
+    return newHead;
 }
